refactor(chat): migrate create-thread call sites to ThreadManagerStore
Swap the four remaining .create.mutateAsync({ data: ... }) callers over
to the new useThreadActions hook from store/hooks. The new create takes
the inner ThreadCreateData shape directly and returns Promise<Task>.
