+++ apps/mesh/src/web/layouts/shell-layout.tsx
@@ -21,7 +21,7 @@ import {
 } from "@tanstack/react-router";
 import { KEYS } from "../lib/query-keys";
 import { readCachedTaskBranch } from "../lib/read-cached-task-branch";
-import { useThreadActions } from "@/web/components/chat/task";
+import { useThreadActions } from "@/web/components/chat/store/hooks";
 import { useOrganizationSettingsSuspense } from "../hooks/use-organization-settings";
 import { useOrgSsoStatus } from "../hooks/use-org-sso";
 import { SsoRequiredScreen } from "../components/sso-required-screen";
@@ -71,7 +71,7 @@ function ShellProjectProvider({
 export function usePanelActions() {
   const navigate = useNavigate();
   const queryClient = useQueryClient();
-  const taskActions = useThreadActions();
+  const { create } = useThreadActions();
   const { org, locator } = useProjectContext();
 
   const params = useParams({ strict: false }) as {
@@ -135,7 +135,7 @@ export function usePanelActions() {
     const targetVmcp =
       search.virtualmcpid ?? getWellKnownDecopilotVirtualMCP(org.id).id;
     try {
-      await taskActions.create.mutateAsync({
+      await create({
         id: newId,
         virtual_mcp_id: targetVmcp,
         ...(branch ? { branch } : {}),
